@@ -1,31 +1,43 @@
-source 'http://rubygems.org'
+source 'https://rubygems.org'
 
-gem 'rails', '3.0.20'
 
-# Bundle edge Rails instead:
-# gem 'rails', :git => 'git://github.com/rails/rails.git'
+gem 'autoprefixer-rails'
+gem 'bootstrap-sass'
+gem 'font-awesome-rails'
+gem 'haml'
+gem 'jquery-rails'
+gem 'pg'
+gem 'rails', '~>4.2.1'
+gem 'sass-rails'
+gem "simple_form"
+gem 'turbolinks'
+gem 'uglifier'
+gem 'httparty'
+gem 'cocoon'
+gem 'bcrypt', '~> 3.1.5', require: "bcrypt"
 
-gem 'sqlite3'
 
-# Use unicorn as the web server
-# gem 'unicorn'
+group :development do
+  gem 'better_errors'
+  gem 'binding_of_caller'
+  gem 'quiet_assets'
+end
 
-# Deploy with Capistrano
-# gem 'capistrano'
+group :test, :development do
+  gem 'database_cleaner'
+  gem 'launchy'
+  gem 'letter_opener'
+  gem 'pry'
+  gem 'rspec-rails'
+end
 
-# To use debugger (ruby-debug for Ruby 1.8.7+, ruby-debug19 for Ruby 1.9.2+)
-# gem 'ruby-debug'
-# gem 'ruby-debug19', :require => 'ruby-debug'
+group :test do
+  gem 'capybara'
+  gem 'factory_girl_rails'
+  gem 'faker'
+  gem 'shoulda-matchers'
+end
 
-# Bundle the extra gems:
-# gem 'bj'
-# gem 'nokogiri'
-# gem 'sqlite3-ruby', :require => 'sqlite3'
-# gem 'aws-s3', :require => 'aws/s3'
-
-# Bundle gems for the local environment. Make sure to
-# put test-only gems in this group so their generators
-# and rake tasks are available in development mode:
-# group :development, :test do
-#   gem 'webrat'
-# end
+group :production do
+  gem 'rails_12factor'
+end
